Deleted test file
Deleted the test file to know whether git will ask for the tokens or not.
I used this line 'git remote set-url origin https://<token>@github.com/<username>/<repo>'.
I will not re-commit again if it works, meaning if I use the above line;
Will it going to ask me for the tokens or not?